[Task 44] Fixed unit tests.
