--- api-core/src/test/java/com/illiapinchuk/moodle/service/impl/TaskServiceImplTest.java
@@ -1,192 +1,0 @@
-package com.illiapinchuk.moodle.service.impl;
-
-import static com.illiapinchuk.moodle.common.TestConstants.TaskConstants.VALID_TASK_1;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import com.illiapinchuk.moodle.common.TestConstants;
-import com.illiapinchuk.moodle.common.mapper.SubmissionMapper;
-import com.illiapinchuk.moodle.common.mapper.TaskMapper;
-import com.illiapinchuk.moodle.common.validator.CourseValidator;
-import com.illiapinchuk.moodle.common.validator.TaskValidator;
-import com.illiapinchuk.moodle.configuration.security.UserPermissionService;
-import com.illiapinchuk.moodle.exception.TaskNotFoundException;
-import com.illiapinchuk.moodle.exception.UserDontHaveAccessToResource;
-import com.illiapinchuk.moodle.persistence.entity.TaskAttachment;
-import com.illiapinchuk.moodle.persistence.repository.TaskRepository;
-import com.illiapinchuk.moodle.service.business.FileUploadService;
-import com.illiapinchuk.moodle.service.business.TaskAttachmentService;
-import java.util.ArrayList;
-import java.util.Optional;
-
-import com.illiapinchuk.moodle.service.impl.business.TaskServiceImpl;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockedStatic;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.web.multipart.MultipartFile;
-
-@ExtendWith(MockitoExtension.class)
-class TaskServiceImplTest {
-
-  @Mock private TaskRepository taskRepository;
-  @Mock private TaskMapper taskMapper;
-  @Mock private TaskValidator taskValidator;
-  @Mock private CourseValidator courseValidator;
-  @Mock private TaskAttachmentService taskAttachmentService;
-  @Mock private FileUploadService fileUploadService;
-  @Mock private SubmissionMapper submissionMapper;
-  @Mock private TaskAttachment taskAttachment;
-  @Mock private MultipartFile file;
-  @InjectMocks private TaskServiceImpl taskService;
-
-  private static MockedStatic<UserPermissionService> mockedUserPermissionService;
-
-  @BeforeAll
-  static void setupUserPermissionServiceMocks() {
-    mockedUserPermissionService = mockStatic(UserPermissionService.class);
-    mockedUserPermissionService
-        .when(UserPermissionService::getJwtUser)
-        .thenReturn(TestConstants.UserConstants.ADMIN_JWT_USER);
-    mockedUserPermissionService.when(UserPermissionService::hasAnyRulingRole).thenReturn(true);
-  }
-
-  @AfterAll
-  static void closeUserPermissionServiceMocks() {
-    mockedUserPermissionService.close();
-  }
-
-  @Test
-  void testAddAttachmentToTask_TaskNotFound() {
-    when(taskRepository.findById(anyString())).thenReturn(Optional.empty());
-
-    assertThrows(
-        TaskNotFoundException.class, () -> taskService.addAttachmentToTask(file, "invalidTaskId"));
-  }
-
-  @Test
-  void testGetTaskById_ValidId_TaskFound() {
-    when(taskRepository.findById(TestConstants.TaskConstants.TASK_ID))
-        .thenReturn(Optional.of(VALID_TASK_1));
-    when(taskAttachmentService.getAttachmentsByTaskId(TestConstants.TaskConstants.TASK_ID))
-        .thenReturn(new ArrayList<>());
-
-    final var result = taskService.getTaskById(TestConstants.TaskConstants.TASK_ID);
-
-    assertEquals(VALID_TASK_1, result);
-    verify(taskAttachmentService).getAttachmentsByTaskId(TestConstants.TaskConstants.TASK_ID);
-  }
-
-  @Test
-  void testGetTaskById_InvalidId_TaskNotFound() {
-    when(taskRepository.findById(TestConstants.TaskConstants.TASK_ID)).thenReturn(Optional.empty());
-
-    assertThrows(
-        TaskNotFoundException.class,
-        () -> taskService.getTaskById(TestConstants.TaskConstants.TASK_ID));
-  }
-
-  @Test
-  void testGetTaskById_UserNotEnrolledWithRulingRole_TaskFound() {
-    when(taskRepository.findById(TestConstants.TaskConstants.TASK_ID))
-        .thenReturn(Optional.of(VALID_TASK_1));
-    when(courseValidator.isStudentEnrolledInCourseWithTask(
-            eq(TestConstants.TaskConstants.TASK_ID), any()))
-        .thenReturn(false);
-    when(taskAttachmentService.getAttachmentsByTaskId(TestConstants.TaskConstants.TASK_ID))
-        .thenReturn(new ArrayList<>());
-
-    final var result = taskService.getTaskById(TestConstants.TaskConstants.TASK_ID);
-
-    assertEquals(VALID_TASK_1, result);
-    verify(taskAttachmentService).getAttachmentsByTaskId(TestConstants.TaskConstants.TASK_ID);
-  }
-
-  @Test
-  void testGetTaskById_UserNotEnrolledWithoutRulingRole_TaskNotFound() {
-    mockedUserPermissionService.when(UserPermissionService::hasAnyRulingRole).thenReturn(false);
-    when(courseValidator.isStudentEnrolledInCourseWithTask(
-            eq(TestConstants.TaskConstants.TASK_ID), any()))
-        .thenReturn(false);
-
-    assertThrows(
-        UserDontHaveAccessToResource.class,
-        () -> taskService.getTaskById(TestConstants.TaskConstants.TASK_ID));
-  }
-
-  @Test
-  void testCreateTask_ValidAuthorId_TaskCreated() {
-    when(taskRepository.save(VALID_TASK_1)).thenReturn(VALID_TASK_1);
-
-    final var result = taskService.createTask(VALID_TASK_1);
-
-    assertEquals(VALID_TASK_1, result);
-    verify(taskRepository).save(VALID_TASK_1);
-  }
-
-  @Test
-  void testUpdateTask_ValidTaskDto_TaskUpdated() {
-    when(taskValidator.isTaskExistsInDbById(TestConstants.TaskConstants.VALID_TASK_DTO.getId()))
-        .thenReturn(true);
-    when(taskRepository.findById(TestConstants.TaskConstants.VALID_TASK_DTO.getId()))
-        .thenReturn(Optional.of(VALID_TASK_1));
-    when(taskRepository.save(VALID_TASK_1)).thenReturn(VALID_TASK_1);
-
-    final var result = taskService.updateTask(TestConstants.TaskConstants.VALID_TASK_DTO);
-
-    assertEquals(VALID_TASK_1, result);
-    verify(taskValidator).isTaskExistsInDbById(TestConstants.TaskConstants.VALID_TASK_DTO.getId());
-    verify(taskRepository).findById(TestConstants.TaskConstants.VALID_TASK_DTO.getId());
-    verify(taskMapper).updateTask(VALID_TASK_1, TestConstants.TaskConstants.VALID_TASK_DTO);
-    verify(taskRepository).save(VALID_TASK_1);
-  }
-
-  @Test
-  void testUpdateTask_InvalidTaskDto_TaskNotFound() {
-    when(taskValidator.isTaskExistsInDbById(TestConstants.TaskConstants.VALID_TASK_DTO.getId()))
-        .thenReturn(false);
-
-    assertThrows(
-        TaskNotFoundException.class,
-        () -> taskService.updateTask(TestConstants.TaskConstants.VALID_TASK_DTO));
-    verify(taskValidator).isTaskExistsInDbById(TestConstants.TaskConstants.VALID_TASK_DTO.getId());
-  }
-
-  @Test
-  void testDeleteTaskById_ValidId_TaskDeleted() {
-    when(taskAttachmentService.getAttachmentsByTaskId(TestConstants.TaskConstants.TASK_ID))
-        .thenReturn(TestConstants.TaskAttachmentConstants.LIST_OF_VALID_ATTACHMENTS);
-
-    taskService.deleteTaskById(TestConstants.TaskConstants.TASK_ID);
-
-    verify(taskAttachmentService).getAttachmentsByTaskId(TestConstants.TaskConstants.TASK_ID);
-    verify(
-            taskAttachmentService,
-            times(TestConstants.TaskAttachmentConstants.LIST_OF_VALID_ATTACHMENTS.size()))
-        .deleteTaskAttachment(any(TaskAttachment.class));
-    verify(taskRepository).deleteById(TestConstants.TaskConstants.TASK_ID);
-  }
-
-  @Test
-  void testGetTasksByCourseId_ValidCourseId_TasksFound() {
-    when(taskRepository.getTasksByCourseId(TestConstants.CourseConstants.VALID_COURSE_ID))
-        .thenReturn(TestConstants.TaskConstants.LIST_OF_VALID_TASKS);
-
-    final var result =
-        taskService.getTasksByCourseId(TestConstants.CourseConstants.VALID_COURSE_ID);
-
-    assertEquals(TestConstants.TaskConstants.LIST_OF_VALID_TASKS, result);
-    verify(taskRepository).getTasksByCourseId(TestConstants.CourseConstants.VALID_COURSE_ID);
-  }
-}
