+++ api-core/src/test/java/com/illiapinchuk/moodle/service/impl/business/EmailServiceImplTest.java
@@ -1,7 +1,6 @@
-package com.illiapinchuk.moodle.service.impl;
+package com.illiapinchuk.moodle.service.impl.business;
 
 import com.illiapinchuk.moodle.common.TestConstants;
-import com.illiapinchuk.moodle.service.impl.business.EmailServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
