+++ api-core/src/test/java/com/illiapinchuk/moodle/service/impl/business/S3FileUploadServiceImplTest.java
@@ -1,4 +1,4 @@
-package com.illiapinchuk.moodle.service.impl;
+package com.illiapinchuk.moodle.service.impl.business;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -14,7 +14,6 @@ import com.illiapinchuk.moodle.exception.CannotWriteToS3Exception;
 import java.io.IOException;
 import java.io.OutputStream;
 
-import com.illiapinchuk.moodle.service.impl.business.S3FileUploadServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
