+++ api-core/src/test/java/com/illiapinchuk/moodle/service/impl/business/TaskSubmissionFacadeImplTest.java
@@ -1,0 +1,5 @@
+package com.illiapinchuk.moodle.service.impl.business;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TaskSubmissionFacadeImplTest {}
