+++ api-core/src/test/java/com/illiapinchuk/moodle/service/impl/business/UserServiceImplTest.java
@@ -1,4 +1,4 @@
-package com.illiapinchuk.moodle.service.impl;
+package com.illiapinchuk.moodle.service.impl.business;
 
 import com.illiapinchuk.moodle.common.TestConstants;
 import com.illiapinchuk.moodle.common.mapper.UserMapper;
@@ -9,7 +9,6 @@ import com.illiapinchuk.moodle.exception.UserNotFoundException;
 import com.illiapinchuk.moodle.model.dto.UserDto;
 import com.illiapinchuk.moodle.persistence.entity.User;
 import com.illiapinchuk.moodle.persistence.repository.UserRepository;
-import com.illiapinchuk.moodle.service.impl.business.UserServiceImpl;
 import jakarta.persistence.EntityExistsException;
 import java.util.Optional;
